# docs/swagger.yaml
basePath: /api/v1
definitions:
  lib.Page:
    properties:
      first:
        description: indicate first value
        example: true
        type: boolean
      items:
        type: object
      last:
        description: indicate last value
        example: false
        type: boolean
      max_page:
        description: maximum pages for current schema
        example: 9
        type: integer
      page:
        description: current page, start from zero
        example: 0
        type: integer
      size:
        description: size per page, default `10`
        example: 10
        type: integer
      total:
        description: total data without filters
        example: 100
        type: integer
      total_pages:
        description: total pages
        example: 10
        type: integer
      visible:
        description: current length
        example: 10
        type: integer
    type: object
  lib.Response:
    properties:
      message:
        description: response message
        type: string
      status:
        description: http status
        type: integer
    type: object
  model.Cart:
    properties:
      additional:
        type: string
      expires_in:
        type: integer
      price:
        type: number
      quantity:
        type: integer
      seat_ids:
        type: string
      seats:
        items:
          $ref: '#/definitions/model.Seat'
        type: array
      show_schedule:
        $ref: '#/definitions/model.ShowSchedule'
      show_schedule_id:
        description: UserID
        format: uuid
        type: string
      total_price:
        type: number
      user_id:
        description: UserID
        format: uuid
        type: string
    type: object
  model.CartAPI:
    properties:
      expires_in:
        type: integer
      price:
        type: number
      quantity:
        type: integer
      seat_ids:
        type: string
      show_schedule_id:
        description: UserID
        format: uuid
        type: string
      total_price:
        type: number
      user_id:
        description: UserID
        format: uuid
        type: string
    type: object
  model.CartPayload:
    properties:
      seatIDs:
        items:
          type: string
        type: array
      showScheduleID:
        type: string
    type: object
  model.ChangePasswordAPI:
    properties:
      confirm_password:
        example: '@Password123'
        type: string
      email:
        example: armadamuhammads@gmail.com
        type: string
      password:
        example: '@Password123'
        minLength: 8
        type: string
    required:
    - confirm_password
    - email
    - password
    type: object
  model.CinemaLocation:
    properties:
      additional:
        type: string
      address:
        type: string
      code:
        type: string
      name:
        type: string
    type: object
  model.LoginAPI:
    properties:
      password:
        example: password
        type: string
      remember:
        example: true
        type: boolean
      username:
        example: armadamuhammads@gmail.com
        type: string
    required:
    - password
    - username
    type: object
  model.LoginResponse:
    properties:
      token:
        $ref: '#/definitions/model.ResponseToken'
      user:
        $ref: '#/definitions/model.User'
    type: object
  model.Movie:
    properties:
      additional:
        type: string
      category:
        type: string
      code:
        type: string
      director:
        type: string
      duration:
        type: number
      name:
        type: string
      producer:
        type: string
    type: object
  model.RegistrationAPI:
    properties:
      confirm_password:
        example: '@Password123'
        type: string
      email:
        example: armadamuhammads@gmail.com
        type: string
      first_name:
        example: Armada
        type: string
      last_name:
        example: Muhammad
        type: string
      password:
        example: '@Password123'
        minLength: 8
        type: string
      phone_number:
        example: "089678009400"
        type: string
      username:
        example: armadamuhammads
        type: string
    required:
    - confirm_password
    - email
    - first_name
    - password
    type: object
  model.ResponseToken:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      is_admin:
        type: boolean
      token_type:
        type: string
    type: object
  model.Seat:
    properties:
      additional:
        type: string
      characteristic:
        type: string
      code:
        type: string
      column:
        type: string
      is_available:
        type: boolean
      row:
        type: string
      seat_layout_id:
        type: string
      show_schedule_id:
        type: string
    type: object
  model.ShowSchedule:
    properties:
      additional:
        type: string
      cinema_location:
        $ref: '#/definitions/model.CinemaLocation'
      cinema_location_id:
        type: string
      date:
        format: date
        type: string
      is_start:
        type: boolean
      movie:
        $ref: '#/definitions/model.Movie'
      movie_id:
        type: string
      price:
        type: number
      start_time:
        format: date-time
        type: string
      theater:
        $ref: '#/definitions/model.Theater'
      theater_id:
        type: string
    type: object
...
  model.Theater:
    properties:
      additional:
        type: string
      cinema_location_id:
        type: string
      code:
        type: string
      name:
        type: string
    type: object
...
  model.User:
    properties:
      additional:
        type: string
      email:
        description: Email
        example: armadamuhammads@gmail.com
        type: string
      first_name:
        description: FirstName
        example: Armada
        type: string
      is_admin:
        description: IsAdmin
        type: boolean
      last_login:
        description: LastLogin
        format: date-time
        type: string
      last_name:
        description: LastName
        example: Muhammad Siswanto
        type: string
      phone_number:
        description: PhoneNumber
        example: "089678009400"
        type: string
      username:
        description: Username
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: armadamuhammads@gmail.com
    name: Armada Muhammad Siswanto
  description: API Documentation
  title: MKP Cinema Ticketing
  version: 1.0.0
paths:
...
  /accounts/change-password:
    post:
      consumes:
      - application/json
      description: Change password
      parameters:
      - description: Account data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.ChangePasswordAPI'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - TokenKey: []
      summary: Change password
      tags:
      - Account
  /accounts/login:
    post:
      consumes:
      - application/json
      description: Login into Account
      parameters:
      - description: Account data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.LoginAPI'
      produces:
      - application/json
      responses:
        "201":
          description: Account data
          schema:
            $ref: '#/definitions/model.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - TokenKey: []
      summary: Login into Account
      tags:
      - Account
  /accounts/register:
    post:
      consumes:
      - application/json
      description: Register new Account
      parameters:
      - description: Account data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.RegistrationAPI'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - TokenKey: []
      summary: Register new Account
      tags:
      - Account
  /carts:
    get:
      consumes:
      - application/json
      description: List of Cart
      parameters:
      - description: Page number start from zero
        in: query
        name: page
        type: integer
      - description: Size per page, default `0`
        in: query
        name: size
        type: integer
      - description: Sort by field, adding dash (`-`) at the beginning means descending
          and vice versa
        in: query
        name: sort
        type: string
      - description: Select specific fields with comma separated
        in: query
        name: fields
        type: string
      - description: custom filters, see [more details](https://github.com/morkid/paginate#filter-format)
        in: query
        name: filters
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of Cart
          schema:
            allOf:
            - $ref: '#/definitions/lib.Page'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/model.Cart'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: List of Cart
      tags:
      - Cart
    post:
      consumes:
      - application/json
      description: Create new Cart
      parameters:
      - description: Cart data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CartPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Cart data
          schema:
            $ref: '#/definitions/model.Cart'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Create new Cart
      tags:
      - Cart
  /carts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Cart by id
      parameters:
      - description: Cart ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete Cart by id
      tags:
      - Cart
    get:
      consumes:
      - application/json
      description: Get a Cart by id
      parameters:
      - description: Cart ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cart data
          schema:
            $ref: '#/definitions/model.Cart'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a Cart by id
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: Update Cart by id
      parameters:
      - description: Cart ID
        in: path
        name: id
        required: true
        type: string
      - description: Cart data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CartAPI'
      produces:
      - application/json
      responses:
        "200":
          description: Cart data
          schema:
            $ref: '#/definitions/model.Cart'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Update Cart by id
      tags:
      - Cart
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  TokenKey:
    in: header
    name: x-Token
    type: apiKey
swagger: "2.0"